+++ experiments/bjoern/exp_model_comparison.py
@@ -8,10 +8,10 @@ import pandas as pd
 import plotly.express as px
 from sklearn.decomposition import PCA
 
+from data_tools.default_paths import default_embeddings_path
+from data_tools.default_paths import default_jsonl_path
 from data_tools.dho_categories import DhOCategory
 from data_tools.message_db import MessageDB
-from experiments.utils.paths import default_embeddings_path
-from experiments.utils.paths import default_jsonl_path
 
 
 def main():
