+++ experiments/experiment_template.py
@@ -1,7 +1,7 @@
 import plotly.express as px
 
+from data_tools.default_paths import default_jsonl_path
 from data_tools.message_db import MessageDB
-from experiments.utils.paths import default_jsonl_path
 
 
 def main():
